requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.14
pydantic==2.5.0
email-validator==2.2.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0
pgvector==0.2.4
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.3
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1 